Feat: Show missing permissions banner earlier in the flow
