+++ src/components/v5/common/ActionSidebar/hooks/useHasActionPermissions/index.ts
@@ -1,0 +1,1 @@
+export * from './useHasActionPermissions.ts';
